Some more list manipulation methods, along with some fixes where we were pushing Ints on to the stack (argh!).

+++ src/main/scala/com/github/calumleslie/hugs/stacks/lib/Maths.scala
@@ -87,7 +87,7 @@ object Maths {
       Consumes an integer 'n', pushes an integer in the range [0,n). n must be <= Integer.MAX_VALUE.
       """, { state: State =>
     state.stack match {
-      case (n: Long) :: rest => state.withStack(random.nextInt(n.toInt) :: rest)
+      case (n: Long) :: rest => state.withStack(random.nextInt(n.toInt).toLong :: rest)
       case _ => throw new IllegalArgumentException(s"Can't apply word to stack ${state.stack}")
     }
   })
